feat: allow anti conceal ignore values to be list of modes
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/204

Updates the recently added `anti_conceal.ignore` settings to allow
values to be a string list of modes in addition to just booleans. This
behaves similarly to the top level `render_modes` setting but at a much
more granular, component level.

It is also a little confusing parsing all the negatives happening,
anti conceal & ignore, what does this mean.

When a list of modes is specified these are the modes where the
decorations will stick around, i.e. not be hidden when your cursor
enters the line. This is similar to neovim's concealcursor behavior.