fix: disable rendering when left column is non 0
## Details

There was exsting logic meant to disable rendering content when the
user scrolls to the right. This is because several elements would need
to alter their behavior based on the column and we would need a complicated
mechanism to handle hidden elements unloading causing misalignment.
Rather than add all this complication for the relatively small benefit
I decided that simply not rendering anything was a better approach.

The logic for this was gated by: `not util.view(win).leftcol == 0`

This does not work because of how the `not` gets assigned:

- What I though: `not (util.view(win).leftcol == 0)`
- Reality: `(not util.view(win).leftcol) == 0`

This would effectively become: `false == 0`, which would never evaluate
to `true`, woops.

The fix is to simply use `~= 0` instead.