fix: reset options when buffer is hidden
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/209

Currently when a user leaves a buffer we render the buffer.

This works as expected when a user is say switching between multiple windows.

However if a user re-uses the current window for another buffer it will
have the window options applied as if it is rendered markdown, i.e. a
conceallevel of 3, which is unepected. This can also occur if a plugin
opens a buffer in the window via the `nvim_win_set_buf` API.

This occurs because at the time the update is triggered the buffer and
window are associated with each other, however due to the use of
vim.schedule when the update occurs this may not longer be the case.

To fix this add a check that the buffer is the one associated with the
window being updated. Include this in the `util.valid` logic so it will
be used at the time the update is triggered and at the time it is
actually ran.

Other minor changes:

- Remove `buf` field from `BufferState`
- Updated window related utility functions
- Remove wrapper functions where they are not needed, i.e. all the
  renderers will work out of the box no need to add one in each module