fix: use first window buffer if buffer is not current window
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/210

Sometimes the buffers associated with windows are switched from another
window. An example of this is the ChatGPT plugin in the issue above
where selecting a session changes the contents of the window, but from
the sessions window.

To get around this rather than always using the current window assuming
events are triggered directly where rendering is needed use the complete
list of windows associated with a buffer and only use the current window
if it is contained in that list.

This is part of the fix to enable this, another requires the plugin to
trigger the TextChanged event so we can update what's rendered.