feat: allow bullet icons to be a list of lists
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/discussions/217

This allows each icon for a bullet to itself be a list of strings. This
list is indexed into using the sibling level of the list item. Sibling
level just means which item number is this in the current list. So you
can do something like roman numerals for level 1 and the alphabet for
level 2. The list is accessed via a clamp so if we exceed the end we
keep using the last value rather than cycling back to the front.

Unrelated change, add a screenshot next to the gifs, see how that looks,
might remove it.