feat: Support disabling background for all languages but keep borders
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/220

This allows the existing `code.disable_background` option to now be a
boolean in addition to a list of languages. Booleans apply the same
behavior for all languages so when set to `true` all languages will have
their backgrounds disabled.

This does alter the behavior of a disabled background, now borders
above and below the code block will continue to be rendered.

If anyone wants to disable the border as well as the background I can
add another flag for that. Decoupling these 2 concepts matches how we
handle headings as well.

Other changes:

- Window options are now set for all windows attached to the buffer
- Adds `padding_text` method to Base to share common pattern
- Removes most usage of window width, replace with number of columns