chore: add screenstring information to all unit tests
## Details

Using screenstring for each row & column we can get a much better idea
of how things actually look to users. Store this in an array of lines
and compare to an expected value.

Keep the existing tests using extmarks since those validate additional
information like highlights which we cannot get from the grid alone.

The only strange thing is we need to trigger a redraw since we're
running in headless mode which does not draw anything by default.

Does take longer for the tests to run, but still pretty fast in general.