feat: roll own type validation, removes reliance on vim.validate
## Details

With the custom validator script we already had most things in place to
remove the underlying usage of vim.validate.

The main part of this change is in better handling of nested types and
capturing non table values where tables are expected.

Since vim.validate is changing in future releases of neovim this avoids
any future required changes to support later versions.