chore(refactor): get config in mark list, configure context update
## Details

In list of marks fetch config from state module directly rather than
having it passed as a parameter. Add a parameter of whether to update
context, which we should do in all handlers except for the main markdown
handler. This avoids adding offset and conceal values that we can't
practically use.

Minor other changes:

- Use assert calls from luassert directly when needed
- Inline require ui to get namespace in test utils