fix: highlight index width calculation
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/212

Calculating the end index of highlighted ranges was done based on
display width but should be done using byte width. Fix is simple and
replaces the usage of `Str.width` with the Lua `#`. Need a better way to
keep the concepts of displayed width and byte width and when to use
which straight. For the most part displayed width is the correct thing
but when placing marks it is not.