feat: render footnote text using superscripts
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/241

When the contents of a link starts with the ^ symbol we'll now assume
this is related to a footnote. When the footnote pattern is spotted we
replace the link contents with a superscript equivalent when possible.
There are some interesting gaps in the superscript glyphs, I do not know
why they exist, but they exist for parens, 0-9, most of a-z, and also
most of A-Z but both character sets have at least one missing glyph.

This behavior can be disabled by setting `link.footnote.superscript` to `false`.

Since the translation between base character and superscript is not
based on a shared offset or anything easy like that we maintain a
complete mapping of base character to superscript. If there are any
missing characters in the mapping then we skip doing any sort of
superscript rendering.