feat: support bare URLs in angle brackets
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/244

Bare URLs in angle brackets are parsed by treesitter into a uri_autolink
node. Add support for this node type in our link renderer. Use the icon
based on the destination. Functions similarly to emails in that the
angle brackets are hidden while the link contents are still displayed
with the addition of the icon.