feat: add text option to html comment concealing
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/244

Previously support was added for html comment concealing gated by the
`html.conceal_comments` option. A request to add some kind of icon next
to the now concealed text was made so I'm breaking this configuration,
should be very minimal impact and having an older version won't fail.

Now we have an `html.comment` configuration which stores a `conceal`
boolean and functions like the previous option. Additionally an optional
`text` field is added along with a `highlight` to use for this text.

By default the `text` field is made nil, so we don't do anything more
by default than before. If a user specifies some string we inline this
at the start of the now concealed comment with the `highlight` specified
or the default which naturally links to `@comment`. Since this is inline
text and not a conceal character users are free to make it arbitrarily
long or use a single icon, all should work without a problem.

Refactored HTML handler to match markdown & markdown_inline approach.