feat: add default icons for ordered lists
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/250

Previously we supported adding icons for ordered lists in much the same
way as unordered, however by default the value was empty.

To do an auto ordering for the pattern of using `1.` for every value we
now provide a default value. However, because ordered lists can go on
and we want to support arbitrary lengths we now also support setting a
function for both ordered and unordered list icons. The function is
called with both the level of the list and the index of the item and
whatever value is returned is used as the icon.

The default function just concatenates a period after the index value
and ignores the level.