feat: add after render callback
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/248

Adds `on.render` callback which is executed after this plugin finishes
rendering a buffer.

This happens after every render cycle, users should be aware that it
will likely be called many times for a given buffer and use it with that
in mind.