feat: only use auto index when marker is 1
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/254

Previously auto indexing was added for ordered lists to generate the
number value based on the items position in the list. This works for
standard lists but sometimes users want to override the number based on
the value actually set. To support this the default value has been
updated to only use the auto value if the user value is <= 1. The icons
function now also takes a third parameter with the text value of the marker.