fix: check buffer is valid when getting name
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/260

When logging info at the buffer level we get the file name for the
buffer to make the logs more useful. However if the buffer is invalid
this call to get file name will throw an exception.

Wrap the call to get file name in a check that buffer is valid. Return
some default string if it is not. Other parts of the ui.update flow will
validate the buffer and window and prevent any actual changes to invalid
buffers.