feat: handle all list marker types in completions
## Details

In completion node type check include all valid unordered list marker
types instead of only the minus marker.

Include the marker characters in the trigger characters list, there will
be a future improvement to add the markers if they are missing but
triggering off of them shouldn't be a problem, though there could be a
missing space.

Use the inline code highlight for padding instead of the normal padding.