feat: improve wiki link rendering
## Details

Discussion: https://github.com/MeanderingProgrammer/render-markdown.nvim/discussions/284

Wiki links are rendered using a single extmark which conceals all
content and then adds inline text as an overlay. This works but means
that any other styles such as a strike through are removed.

To improve this handle each component of the wiki link separately,
similarly to the way autolinks are handled. This means we:

- Hide the opening bracket
- Inline the icon
- Hide the link destination if an alias exists
- Hide the closing bracket

As 4 separate individual marks. This means the contents of the alias &
destination will keep the style they had before rendering.