fix: ensure space for cell padding
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/287

We have a default padding of 1 set for pipe table cells. However when
there is not actually enough space available to create this padding we
end up with unhandled negative values which lead to poor rendering.

To fix this adjust column width calculation to account for needed space.

For example if we have a cell like `|foo|`, it would normally have a
width of 3. Naturally there is not enough space on either side for
padding. The fix adds to the width value for each cell to ensure it can
allow for the specified amount of padding on both side, so now the width
calculation is 5. With a column width of 5 and the cell still having an
actual width of 3 we'll end up doing the correct thing and adding a
space on both sides resulting in `| foo |`.