feat(completions): insert space after marker
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/292

Updates the logic of completion text to check if the marker node (for
both list items and block quotes) ends in a space. If it does not
then include a space in the completions.

Also adds a call to treesitter parser `parse` method for current row
to remove various edge cases with treesitter not being up to date after
text is inserted depending on when completion plugin runs each
completion source. Should not have a noticeable performance impact since
we limit range to current row.