fix: bullet padding indenting heading borders
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/297

Tree sitter parser has the last list item extending to the next section.
Since heading borders are added as overlay lines when space is available
they will end up getting the padding that was added there for the list
item. Fix is to have a special case for the last line in a list when it
is empty, not the best but seems to work well enough.