fix: check if highlighter is active before computing concealed regions
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/300

While we report inactive highlights in the healthcheck we may as well
handle the case when highlights are not active accurately. To do this
all we need to do is check if highlights are active for the buffer
before iterating through the highlighter query.