feat: handle possible issue with initial load when lazy loading
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/309

Depending on how the plugin manager implements lazy loading, this plugin can
miss the initial buffer that triggered it to load.

This is because we setup a `FileType` autocommand to check which buffers
we should attach to. If the user configures this plugin to load based on
a filetype that event will not be processed by the autocommand (since it
has not been created yet). Instead everything will be setup but the
current buffer will end up ignored.

I'm unsure why this is not an issue with `lazy.nvim`, maybe it executes
autocommands after the initial load. This would make sense, but I have
not verified it.

To get around this attempt to attach to the current buffer at the start,
similar to the command based lazy loading edge case.