fix: schedule initial buffer attach to handle lazy loading edge cases
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/315

Previously we added a step to attach to the current buffer to handle
`vim-plug` based lazy loading not running for the initial file.

This causes a problem with `lazy.nvim` since the attach function will
run before picking up user configuration values, which will cause the
first buffer to be out of sync from the user configuration.

As a sort of hacky but good enough solution run the lazy load handling
attach calls behind a vim.schedule. This should result in the event
based ones running first and avoid this problem.