refactor!: pass context table to all user functions instead of parameters
## Details

This modifies every user accessible function to take a single context
table as input rather than a growing list of parameters. This makes
adding parameters to existing functions as well as using the functions
easier since you no longer need to pay attention to the order of
arguments but does unfortunately break existing users, hence the major
version version bump.

The fix to any errors is straightforward as the fields of the context
continue to have the same names. So if you have an existing function like:

```lua
function(param1, param2)
    vim.print(param1)
    vim.print(param2)
end
```

The fixed version would be:

```lua
function(ctx)
    vim.print(ctx.param1)
    vim.print(ctx.param2)
end
```

If you use the same parameters many times in the function and don't want
to add the `ctx.` prefix everywhere you can add a line at the top to
define local variables with the same name as before and keep the rest of
the function body unchanged:

```lua
function(ctx)
    local param1, param2 = ctx.param1, ctx.param2
    vim.print(param1)
    vim.print(param2)
end
```

The fields impacted are:

- The `parse` functions in `custom_handlers`:
  - `buf` -> `ctx.buf`
  - `root` -> `ctx.root`
- The callbacks in `on`: `on.attach` & `on.render`:
  - `buf` -> `ctx.buf`
- `bullet.icons` & `bullet.ordered_icons` if a non-default function was set:
  - `level` -> `ctx.level`
  - `index` -> `ctx.index`
  - `value` -> `ctx.value`
- `heading.icons` if a function was set:
  - `sections` -> `ctx.sections`