feat: support custom heading properties based on heading text
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/320

The desired feature was some way to change the icon, foreground, and
background for a heading based on the text of the heading.

To accomplish have added a new configuration under `heading` called
`custom` which functions very similarly to custom links. Users define a
`pattern` which will be matched against the heading text and if it does
we will use the `icon`, `foreground`, & `background` (if provided) from
the matching custom configuration instead of the top level values.

For example:

```lua
require('render-markdown').setup({
    heading = {
        custom = {
            codecompanion_me = {
                pattern = '^## Me$',
                icon = 'some_icon',
                foreground = 'some_highlight',
                background = 'some_highlight',
            },
        },
    },
})
```