feat: set default render_modes to true (all) for LSP docs
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/326

When triggering LSP hover docs / signature help in insert mode this
plugin will attach to the buffer but since we're in insert mode it will
display the raw view instead of the rendered view. By setting the
`render_modes` to `true` now the doc buffers will always be in the
rendered state.