refactor: remove error false in parser health check
## Details

Removes `{ error = false }` from `vim.treesitter.get_parser` call. Not
necessary since the call is wrapped in a `pcall` as is and the parameter
will be removed in `0.12`.

https://neovim.io/doc/user/treesitter.html#vim.treesitter.get_parser()

```
If no parser can be created, an error is thrown. Set opts.error = false
to suppress this and return nil (and an error message) instead. WARNING:
This behavior will become default in Nvim 0.12 and the option will be
removed.
```