feat(0.11): process conceal_lines metadata from highlights
## Details

This is some of the work needed to better handle the recent
conceal_lines change added to neovim:
https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/352

In order to better render code blocks, or at least have the ability to,
we first need to figure out which lines are impacted by the change. To
do this update the `conceal` module to process this additional metadata
and store it. Expose this information again through the `context` module.

This change does not include using this information to actually improve
the rendering, but having it ready to go is useful, and the code block
implementation to get around this is going to be pretty tedious.

Some minor refactoring along the way to make these changes a little easier.