feat: add scope_highlight to list items
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/354

Adds a `scope_highlight` parameter to the `bullet` configuration which
is used in the same way as the `checkbox` equivalent to highlight the item.

Also allows `highlight` to be the same type as `icons`, i.e. `string` |
`string[]` | `string[][]` | `function`. The support for a plain string
was added in this change as well to ensure backwards compatibility for
the `highlight` configuration.

Update documentation formatting and improve validation messages.