fix: lsp hover doc for neovim < 0.11.0
## Details

Follow up to the fix for 0.11.0:
https://github.com/MeanderingProgrammer/render-markdown.nvim/commit/b57d51d760f5e4f520414dbeb6dee3ec1ae07a83

That change ended up causing the same problem that existed for 0.11.0
for older versions of neovim, i.e. not rendering on the initial load.

This is due to when the LSP window is created. On the `initial` run the
first call would use a non-final floating window. The final one gets
created later. Due to the debounce the next call with the correct
window was skipped.

To fix this add a check that the buffer is not empty before triggering
the update. This should work for all versions and skipping empty buffers
should always be valid.

Minor refactor: rename `BufferState` -> `Buffer`.