feat: provide in-process lsp completion source
## Details

Rather than requiring a source to be added to completion engine by user
create an in-process LSP that provides completion capabilities to
whatever engine the user happens to be running.

That way the setup is the same for both `nvim-cmp` & `blink.cmp` and
more minimal since LSPs are enabled by default by completion engines.

Idea for this and the implementation came from [saecki/crates.nvim](https://github.com/saecki/crates.nvim)
and seems to have no downside. Can add more capabilities in the future
though I don't think there's a strong use case for this yet.

To avoid breaking existing `nvim-cmp` users we add opt-in configurations
for `lsp` & `coq`, and if neither of these are set we continue to
register the `nvim-cmp` source, keeping default behavior effectively unchanged.

To enable this feature set:

```lua
require('render-markdown').setup({
    completions = { lsp = { enabled = true } },
})
```

Since having multiple of these sources running at once would result in
duplicate suggestions we only setup the first one that is enabled in the
order `lsp` -> `coq` -> `cmp`.