feat: add stubs for blink.cmp source registration
## Details

Will do the equivalent of adding the markdown source to `blink.cmp` once
the `add_source_provider` API is released.

Currently it is not so just print a warning to any users that attempt to
use it before this release.

Leave the `blink.cmp` integration instructions unchanged until (likely
next) release of `blink.cmp`.