fix: always add indent to column when right aligning code language
## Details

Previously when right aligning code language the amount of indent added
would only be accounted for with the `block` width. This means the
default `full` width would end up drifting the heading more left as the
level increased (if indent mode was enabled that is).

This was not intended behavior, so add the indent regardless of `width`.