feat: improve completion source
## Details

Continue to provide completions once inside of square brackets and not
just when starting a list item or block quote.

Requires handling a generic `paragraph` node which is the type of the
body for both and handling different scenarios for the node. So as not
to provide bad completions do not return anything if the user is typing
outside of the start of the handled node types.

This does mean we no longer have suggestions before starting a marker,
meaning going to the next line in an existing list now returns nothing
as we do not know if you want to continue the list and handling this
behavior well is complicated.

In theory anytime the user goes to a new line outside of few contexts
they could start a new list or a block quote, so empty lines could
always be provided with completions, but IMO this is pretty bad
behavior. So as not to over-complicate things we now always require
intent from the user, meaning starting a list item (via `-`, `*`, or
`+`) or a block quote (via `>`) is what triggers completions, otherwise
do not provide completions.

Create unit tests for completions source to validate behavior.