feat: automatic source registration for blink.cmp
## Details

Follows release of `blink.cmp` 0.14.0, which provides additional APIs
for registering sources.

Now when the `blink` completion source is enabled in this plugin we'll
use the `add_source_provider` & `add_filetype_source` functions to
automatically register the completion source on behalf of the user.

This is an opt-in feature so out of the box nothing new happens.

The in-process LSP source is still the suggested approach for getting
completions from this plugin as they are completion engine agnostic.

Simplify the instructions for using the `blink.cmp` source in the README.